django/requirements.txt
asgiref==3.4.1
Django==3.2.7
django-rest-framework==0.1.0
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
gunicorn==20.1.0
psycopg2-binary==2.9.1
python-dotenv==0.17.1

